+++ src/app/components/Home.js
@@ -11,7 +11,7 @@ import MomentUtils from '@date-io/moment';
 import * as Sentry from '@sentry/react';
 import config from 'config'; // eslint-disable-line require-path-exists/exists
 import LoginContainer from './login/LoginContainer';
-import InviteNewAccount from './InviteNewAccount';
+import InviteNewAccount from './login/InviteNewAccount';
 import BrowserSupport from './BrowserSupport';
 import CheckContext from '../CheckContext';
 import DrawerNavigation from './DrawerNavigation';
@@ -274,7 +274,7 @@ class HomeComponent extends Component {
       return false;
     })();
 
-    if (loggedIn && user.completed_signup) {
+    if (loggedIn && !user.completed_signup) {
       return (
         <InviteNewAccount teamSlug={teamSlug} />
       );
