+++ src/app/components/login/InviteNewAccount.js
@@ -3,39 +3,17 @@ import Relay from 'react-relay/classic';
 import { QueryRenderer, graphql, commitMutation } from 'react-relay/compat';
 import { FormattedMessage, FormattedHTMLMessage } from 'react-intl';
 import PropTypes from 'prop-types';
-import Button from '@material-ui/core/Button';
-import Typography from '@material-ui/core/Typography';
-import { makeStyles } from '@material-ui/core/styles';
 import cx from 'classnames/bind';
-import TextField from './cds/inputs/TextField';
-import Alert from './cds/alerts-and-prompts/Alert';
-import UserTosForm from './UserTosForm';
-import { FormattedGlobalMessage } from './MappedMessage';
-import GenericUnknownErrorMessage from './GenericUnknownErrorMessage';
-import { stringHelper } from '../customHelpers';
-import { getErrorMessageForRelayModernProblem } from '../helpers';
-import styles from './login/login.module.css';
-import inputStyles from '../styles/css/inputs.module.css';
-import {
-  units,
-} from '../styles/js/shared';
-
-const useStyles = makeStyles({
-  logo: {
-    margin: '0 auto',
-    display: 'block',
-  },
-  primaryButton: {
-    display: 'block',
-    margin: `${units(2)} auto`,
-  },
-  bold: {
-    fontWeight: 'bold',
-  },
-  topMargin: {
-    marginTop: `${units(3)}`,
-  },
-});
+import ButtonMain from '../cds/buttons-checkboxes-chips/ButtonMain';
+import TextField from '../cds/inputs/TextField';
+import Alert from '../cds/alerts-and-prompts/Alert';
+import UserTosForm from '../UserTosForm';
+import { FormattedGlobalMessage } from '../MappedMessage';
+import GenericUnknownErrorMessage from '../GenericUnknownErrorMessage';
+import { stringHelper } from '../../customHelpers';
+import { getErrorMessageForRelayModernProblem } from '../../helpers';
+import styles from './login.module.css';
+import inputStyles from '../../styles/css/inputs.module.css';
 
 const InviteNewAccountComponent = ({ user }) => {
   const [message, setMessage] = React.useState(null);
@@ -111,8 +89,6 @@ const InviteNewAccountComponent = ({ user }) => {
     setCheckedTos(!checkedTos);
   };
 
-  const classes = useStyles();
-
   return (
     <div className={styles['login-wrapper']}>
       <div className={styles['login-container']}>
@@ -121,38 +97,36 @@ const InviteNewAccountComponent = ({ user }) => {
             <FormattedGlobalMessage messageKey="appNameHuman">
               {appNameHuman => (
                 <img
+                  className={cx('login__icon', styles['login-logo'])}
                   alt={appNameHuman}
-                  width="120"
-                  className={['login__icon', classes.logo].join(' ')}
                   src={stringHelper('LOGO_URL')}
+                  width="120"
                 />
               )}
             </FormattedGlobalMessage>
-
-            <Typography component="div" align="center" className={classes.topMargin}>
-              <FormattedMessage
-                id="inviteNewAccount.invitedBy"
-                defaultMessage="{name} has invited you to join the workspace"
-                description="Message to the current user about who has invited them to join this workspace"
-                values={{
-                  name: teamUser?.invited_by?.name,
-                }}
-              />
-            </Typography>
-            <Typography component="div" align="center" className={classes.bold} paragraph>
-              {teamUser?.team?.name}
-            </Typography>
-            <Typography component="div" align="center" paragraph>
+            <div className={styles['login-form-invited']}>
+              <p>
+                <FormattedMessage
+                  id="inviteNewAccount.invitedBy"
+                  defaultMessage="{name} has invited you to join the workspace:"
+                  description="Message to the current user about who has invited them to join this workspace"
+                  values={{
+                    name: teamUser?.invited_by?.name,
+                  }}
+                />
+                <br />
+                <strong>{teamUser?.team?.name}</strong>
+              </p>
               <FormattedHTMLMessage
+                tagName="p"
                 id="inviteNewAccount.createMessage"
-                defaultMessage="You need to create an account for <b>{email}</b>"
+                defaultMessage="You need to create an account for <strong>{email}</strong>"
                 values={{
                   email: user.email,
                 }}
                 description="Inform your to create a new account for signup"
               />
-            </Typography>
-
+            </div>
             {message &&
               <Alert
                 className={styles['login-form-alert']}
@@ -244,19 +218,23 @@ const InviteNewAccountComponent = ({ user }) => {
               />
             </div>
             <div className="login__actions">
-              <Button
+              <ButtonMain
+                size="default"
+                theme="brand"
                 variant="contained"
-                color="primary"
-                type="submit"
-                id="submit-register-or-login"
-                className={['login__submit login__submit--register', classes.primaryButton].join(' ')}
-              >
-                <FormattedMessage
-                  id="inviteNewAccount.createAccount"
-                  defaultMessage="Create Account"
-                  description="Submit button for create a new account"
-                />
-              </Button>
+                buttonProps={{
+                  id: 'submit-register-or-login',
+                  type: 'submit',
+                }}
+                className={cx('login__submit', 'login__submit--register')}
+                label={
+                  <FormattedMessage
+                    id="inviteNewAccount.createAccount"
+                    defaultMessage="Create Account"
+                    description="Submit button for create a new account"
+                  />
+                }
+              />
             </div>
           </form>
         </div>
