+++ src/app/components/login/InviteNewAccount.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { mountWithIntlProvider } from '../../../test/unit/helpers/intl-test';
+import { mountWithIntlProvider } from '../../../../test/unit/helpers/intl-test';
 
 import { InviteNewAccountComponent } from './InviteNewAccount';
